+++ core/dbt/contracts/graph/nodes.py
@@ -42,7 +42,8 @@ from dbt.contracts.graph.unparsed import (
 )
 from dbt.contracts.graph.node_args import ModelNodeArgs
 from dbt.contracts.graph.semantic_layer_common import WhereFilterIntersection
-from dbt.contracts.util import Replaceable, AdditionalPropertiesMixin
+from dbt.contracts.util import Replaceable
+from dbt.common.contracts.config.properties import AdditionalPropertiesMixin
 from dbt.common.events.functions import warn_or_error
 from dbt.exceptions import ParsingError, ContractBreakingChangeError
 from dbt.common.events.types import (
