+++ tests/adapter/dbt/tests/adapter/materialized_view/changes.py
@@ -3,7 +3,7 @@ from typing import Optional
 import pytest
 
 from dbt.adapters.base.relation import BaseRelation
-from dbt.contracts.graph.model_config import OnConfigurationChangeOption
+from dbt.common.contracts.config.materialization import OnConfigurationChangeOption
 from dbt.adapters.contracts.relation import RelationType
 from dbt.tests.util import (
     assert_message_in_logs,
